refactor: Move Assembly loading to shared AssemblyReader class in MdDocs.Common
