+++ src/MdDocs.Common/Grynwald.MdDocs.Common.csproj
@@ -7,6 +7,8 @@
 
   <ItemGroup>
     <PackageReference Include="Grynwald.MarkdownGenerator" Version="2.2.14-pre" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="2.2.0" />
+    <PackageReference Include="Mono.Cecil" Version="0.11.0" />
   </ItemGroup>
   
 </Project>
